Make types for offlineStateLens consistent
The default `offlineStateLens` allows the offline state to be `set` to
`null` or `undefined`, but was expecting `get` to return a non-optional
`OfflineState` which is not always the case. Updating the return type to
match the input type better reflects the possible values which can be
returned by the `get` function.

This also caters for the case where the offline state is missing by
logging a warning instead of throwing an error when trying to access
properties from the non-existent offline object.