Add enhanceReplaceReducer config to createOffline
The `createOffline` function is supposed to let the developer decide
where to insert each of `middleware`, `enhanceStore` and
`enhanceReducer` in the code, but the monkey patch always wrapped the next
reducer with `enhanceReducer`.

Disabling the monkey patch allows the new reducer to be wrapped where it
makes sense so that the same flexibility is available when replacing the
reducer as was available for the initial reducer when the store was
created.